# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NIPOPETO Pet Healthcare API",
    "version": "3.0.0",
    "description": "REST/JSON contract with bounded session/dashboard bootstrap and paginated collections. List GET response changed from array to {items,pagination,counts,related}. Default 20, maximum 100 rows."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "paths": {
...
    "/medical-records": {
      "get": {
        "tags": [
          "Medical records"
        ],
        "summary": "List medical records",
        "operationId": "get_api_medical_records",
        "responses": {
          "200": {
            "description": "Page of MedicalRecord",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "pagination",
                    "counts",
                    "related"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MedicalRecord"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "counts": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "integer",
                        "minimum": 0
                      }
                    },
                    "related": {
                      "$ref": "#/components/schemas/RelatedPageData"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Paginated JSON envelope. Rows use a stable order with id as a tie-breaker. Count, filters and page share a Repeatable Read transaction. related contains only relations needed by this page. Counts are independent of the selected status/page.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/SearchQuery"
          },
          {
            "$ref": "#/components/parameters/ExactId"
          },
          {
            "$ref": "#/components/parameters/OwnerIdFilter"
          },
          {
            "$ref": "#/components/parameters/PetIdFilter"
          }
        ]
      },
      "post": {
        "tags": [
          "Medical records"
        ],
        "summary": "Create medical record",
        "operationId": "post_api_medical_records",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicalRecordEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicalRecordInput"
              }
            }
          }
        }
      }
    },
    "/medical-records/{id}": {
      "patch": {
        "tags": [
          "Medical records"
        ],
        "summary": "Update medical record",
        "operationId": "patch_api_medical_records_id",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicalRecordEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicalRecordUpdateInput"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Medical records"
        ],
        "summary": "Delete medical record",
        "operationId": "delete_api_medical_records_id",
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ]
      }
    },
    "/medical-records/images": {
      "post": {
        "tags": [
          "Medical records"
        ],
        "summary": "Add medical image",
        "operationId": "post_api_medical_records_images",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicalImageEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicalImageInput"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listMedicalImage",
        "summary": "List MedicalImage",
        "tags": [
          "Pagination"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Paginated JSON envelope. Rows use a stable order with id as a tie-breaker. Count, filters and page share a Repeatable Read transaction. related contains only relations needed by this page. Counts are independent of the selected status/page.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/SearchQuery"
          },
          {
            "$ref": "#/components/parameters/ExactId"
          },
          {
            "$ref": "#/components/parameters/PetIdFilter"
          }
        ],
        "responses": {
          "200": {
            "description": "Page of MedicalImage",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "pagination",
                    "counts",
                    "related"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MedicalImage"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "counts": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "integer",
                        "minimum": 0
                      }
                    },
                    "related": {
                      "$ref": "#/components/schemas/RelatedPageData"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/medical-records/images/{id}": {
      "delete": {
        "tags": [
          "Medical records"
        ],
        "summary": "Delete medical image",
        "operationId": "delete_api_medical_records_images_id",
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ]
      }
    },
...
    "/notifications/read-all": {
      "patch": {
        "tags": [
          "Notifications"
        ],
        "summary": "Mark all notifications read",
        "operationId": "patch_api_notifications_read_all",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": " Notification mutations use the same inbox scope as listing: owner recipient ID or the admin-role staff inbox."
      }
    },
    "/notifications/{id}/read": {
      "patch": {
        "tags": [
          "Notifications"
        ],
        "summary": "Mark notification read",
        "operationId": "patch_api_notifications_id_read",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "description": " Notification mutations use the same inbox scope as listing: owner recipient ID or the admin-role staff inbox."
      }
    },
    "/notifications/{id}": {
      "delete": {
        "tags": [
          "Notifications"
        ],
        "summary": "Delete notification",
        "operationId": "delete_api_notifications_id",
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "description": " Notification mutations use the same inbox scope as listing: owner recipient ID or the admin-role staff inbox."
      }
    },
    "/notifications/send": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Send notification",
        "operationId": "post_api_notifications_send",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationSendInput"
              }
            }
          }
        },
        "description": " Notification mutations use the same inbox scope as listing: owner recipient ID or the admin-role staff inbox."
      }
    },
...
  },
  "components": {
    "responses": {
...
      "ApiError": {
        "description": "Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 1000000,
          "default": 1
        },
        "description": "1-based page."
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        },
        "description": "Rows per page; values above 100 return 422."
      },
      "SearchQuery": {
        "name": "q",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 100
        },
        "description": "Case-insensitive search applied in the database before pagination."
      },
      "ExactId": {
        "name": "id",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 128
        },
        "description": "Exact identifier lookup within the authenticated scope."
      },
      "OwnerIdFilter": {
        "name": "ownerId",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 128
        },
        "description": "Staff filter; owners are always scoped to their own JWT subject."
      },
      "ResourceId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "PetIdFilter": {
        "name": "petId",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 128
        },
        "description": "Optional database filter. Use all or omit for no filter."
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Message": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "Pet": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "species": {
            "type": "string"
          },
          "breed": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "ageLabel": {
            "type": "string"
          },
          "weightKg": {
            "type": "number"
          },
          "microchipId": {
            "type": "string"
          },
          "healthStatus": {
            "type": "string"
          },
          "allergies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/String"
            }
          },
          "qrToken": {
            "type": "string"
          },
          "qrEnabled": {
            "type": "boolean"
          },
          "publicProfile": {
            "$ref": "#/components/schemas/PublicProfile"
          }
        },
        "required": [
          "id",
          "ownerId",
          "name",
          "species",
          "gender",
          "healthStatus",
          "allergies",
          "qrEnabled",
          "publicProfile"
        ]
      },
...
      "MedicalRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "petId": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "appointmentId": {
            "type": "string"
          },
          "doctorName": {
            "type": "string"
          },
          "visitDate": {
            "type": "string",
            "format": "date"
          },
          "title": {
            "type": "string"
          },
          "diagnosis": {
            "type": "string"
          },
          "treatment": {
            "type": "string"
          },
          "weightKg": {
            "type": "number"
          },
          "temperatureC": {
            "type": "number"
          },
          "heartRateBpm": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "petId",
          "ownerId",
          "doctorName",
          "visitDate",
          "title",
          "diagnosis",
          "treatment"
        ]
      },
      "MedicalImage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "petId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "petId",
          "title",
          "imageUrl",
          "mimeType"
        ]
      },
...
      "MedicalRecordEnvelope": {
        "type": "object",
        "properties": {
          "record": {
            "$ref": "#/components/schemas/MedicalRecord"
          }
        },
        "required": [
          "record"
        ]
      },
      "MedicalImageEnvelope": {
        "type": "object",
        "properties": {
          "image": {
            "$ref": "#/components/schemas/MedicalImage"
          }
        },
        "required": [
          "image"
        ]
      },
...
      "NotificationEnvelope": {
        "type": "object",
        "properties": {
          "notification": {
            "$ref": "#/components/schemas/Notification"
          }
        },
        "required": [
          "notification"
        ]
      },
...
      "MedicalRecordInput": {
        "type": "object",
        "properties": {
          "petId": {
            "type": "string"
          },
          "appointmentId": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000,
            "description": "Trimmed; whitespace-only values are rejected."
          },
          "visitDate": {
            "type": "string",
            "format": "date",
            "description": "Real calendar date; dates such as February 30 are rejected."
          },
          "doctorName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000,
            "description": "Trimmed; whitespace-only values are rejected."
          },
          "diagnosis": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000,
            "description": "Trimmed; whitespace-only values are rejected."
          },
          "treatment": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000,
            "description": "Trimmed; whitespace-only values are rejected."
          },
          "weightKg": {
            "oneOf": [
              {
                "type": "number",
                "minimum": 0.01,
                "maximum": 999.99,
                "multipleOf": 0.01,
                "nullable": true
              },
              {
                "type": "string",
                "pattern": "^\\d+(?:\\.\\d{1,2}0*)?$",
                "description": "Decimal string; same numeric range applies."
              },
              {
                "type": "string",
                "enum": [
                  ""
                ],
                "description": "Empty string clears the optional value."
              }
            ],
            "description": "Positive optional value with at most 2 decimal places; null or empty string clears it."
          },
          "symptoms": {
            "type": "string",
            "nullable": true,
            "maxLength": 10000
          },
          "medications": {
            "type": "string",
            "nullable": true,
            "maxLength": 10000
          },
          "vaccineName": {
            "type": "string",
            "nullable": true,
            "maxLength": 10000
          },
          "internalNote": {
            "type": "string",
            "nullable": true,
            "maxLength": 10000
          },
          "followUpDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Real calendar date, YYYY-MM-DD; null clears it."
          },
          "temperatureC": {
            "oneOf": [
              {
                "type": "number",
                "minimum": 0.1,
                "maximum": 999.9,
                "multipleOf": 0.1,
                "nullable": true
              },
              {
                "type": "string",
                "pattern": "^\\d+(?:\\.\\d{1,1}0*)?$",
                "description": "Decimal string; same numeric range applies."
              },
              {
                "type": "string",
                "enum": [
                  ""
                ],
                "description": "Empty string clears the optional value."
              }
            ],
            "description": "Positive optional value with at most 1 decimal places; null or empty string clears it."
          },
          "heartRateBpm": {
            "oneOf": [
              {
                "type": "integer",
                "minimum": 0,
                "exclusiveMinimum": true,
                "maximum": 2147483647,
                "nullable": true
              },
              {
                "type": "string",
                "pattern": "^\\d+$",
                "description": "Decimal string; same numeric range applies."
              },
              {
                "type": "string",
                "enum": [
                  ""
                ],
                "description": "Empty string clears the optional value."
              }
            ]
          }
        },
        "required": [
          "petId",
          "title",
          "visitDate",
          "doctorName",
          "diagnosis",
          "treatment"
        ]
      },
      "MedicalImageInput": {
        "type": "object",
        "properties": {
          "petId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          }
        },
        "required": [
          "petId",
          "title",
          "imageUrl",
          "mimeType"
        ]
      },
...
      "NotificationSendInput": {
        "type": "object",
        "properties": {
          "recipientOwnerId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "recipientOwnerId",
          "title",
          "message"
        ]
      },
...
      "MedicalRecordUpdateInput": {
        "type": "object",
        "description": "Staff may edit medical details only. petId, ownerId and appointmentId are fixed after creation; attempts to change or clear them return 422. Older clients may send unchanged values.",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000,
            "description": "Trimmed; whitespace-only values are rejected."
          },
          "visitDate": {
            "type": "string",
            "format": "date",
            "description": "Real calendar date; dates such as February 30 are rejected."
          },
          "doctorName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000,
            "description": "Trimmed; whitespace-only values are rejected."
          },
          "diagnosis": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000,
            "description": "Trimmed; whitespace-only values are rejected."
          },
          "treatment": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000,
            "description": "Trimmed; whitespace-only values are rejected."
          },
          "weightKg": {
            "oneOf": [
              {
                "type": "number",
                "minimum": 0.01,
                "maximum": 999.99,
                "multipleOf": 0.01,
                "nullable": true
              },
              {
                "type": "string",
                "pattern": "^\\d+(?:\\.\\d{1,2}0*)?$",
                "description": "Decimal string; same numeric range applies."
              },
              {
                "type": "string",
                "enum": [
                  ""
                ],
                "description": "Empty string clears the optional value."
              }
            ],
            "description": "Positive optional value with at most 2 decimal places; null or empty string clears it."
          },
          "symptoms": {
            "type": "string",
            "nullable": true,
            "maxLength": 10000
          },
          "medications": {
            "type": "string",
            "nullable": true,
            "maxLength": 10000
          },
          "vaccineName": {
            "type": "string",
            "nullable": true,
            "maxLength": 10000
          },
          "internalNote": {
            "type": "string",
            "nullable": true,
            "maxLength": 10000
          },
          "followUpDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Real calendar date, YYYY-MM-DD; null clears it."
          },
          "temperatureC": {
            "oneOf": [
              {
                "type": "number",
                "minimum": 0.1,
                "maximum": 999.9,
                "multipleOf": 0.1,
                "nullable": true
              },
              {
                "type": "string",
                "pattern": "^\\d+(?:\\.\\d{1,1}0*)?$",
                "description": "Decimal string; same numeric range applies."
              },
              {
                "type": "string",
                "enum": [
                  ""
                ],
                "description": "Empty string clears the optional value."
              }
            ],
            "description": "Positive optional value with at most 1 decimal places; null or empty string clears it."
          },
          "heartRateBpm": {
            "oneOf": [
              {
                "type": "integer",
                "minimum": 0,
                "exclusiveMinimum": true,
                "maximum": 2147483647,
                "nullable": true
              },
              {
                "type": "string",
                "pattern": "^\\d+$",
                "description": "Decimal string; same numeric range applies."
              },
              {
                "type": "string",
                "enum": [
                  ""
                ],
                "description": "Empty string clears the optional value."
              }
            ]
          }
        }
      },
...
      "Owner": {
        "allOf": [
          {
            "$ref": "#/components/schemas/User"
          },
          {
            "type": "object",
            "required": [
              "petIds",
              "loginEnabled"
            ],
            "properties": {
              "petIds": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "loginEnabled": {
                "type": "boolean",
                "description": "True only when this profile has login credentials."
              }
            }
          }
        ],
        "properties": {
          "petCount": {
            "type": "integer",
            "minimum": 0,
            "description": "Total pet count independent of pages. petIds is empty in bounded directory/profile responses; load pet pages with ownerId."
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ],
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "RelatedPageData": {
        "type": "object",
        "properties": {
          "owners": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Owner"
            }
          },
          "pets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Pet"
            }
          },
          "appointments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Appointment"
            }
          },
          "hotelBookings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HotelBooking"
            }
          },
          "medicalRecords": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicalRecord"
            }
          },
          "medicalImages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicalImage"
            }
          },
          "dailyCareNotes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CareNote"
            }
          },
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            }
          },
          "invoices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Invoice"
            }
          }
        }
      }
    }
  }
}